feat(db): add CreateKeySpace function
Add a new function called CreateKeySpace to the db package in the
ScyllaDB initialization module. This function creates a keyspace with
the given name using the provided session and executes a CQL statement
to create the keyspace if it doesn't already exist. The replication
strategy used is 'SimpleStrategy' with a replication factor of 1.

Issue: #feature/source-repository