feat(scylladb): Add package for initializing a connection to ScyllaDB database
Added a new file `db.go` in the `source/adapter/scylladb` directory.
This file contains the `scylladb` package, which provides functionality
for initializing a connection to ScyllaDB database using the `gocql`
library.

The package defines a configuration structure (`Config`) and a function
(`New`) to create and return a session interface for interacting with
ScyllaDB. The `New` function takes a `Config` parameter to customize the
behavior of the ScyllaDB connection.